Update CI workflow for Arduino project-specific linting
On every push, pull request, and periodically, run Arduino Lint to check for common problems not related to the project
code.

This replaces the "Extra Library Checks" workflow in all but one respect: keywords.txt linting. Arduino Lint does not
currently do any validation of the keywords.txt data, so use of the `per1234/arduino-ci-script` script is still required
for this specific check. A dedicated "Check keywords.txt" workflow was added for that purpose.